Cleaned up main code and modified reaction diffusion simulation to use web workers
