+++ src/App.css
@@ -1,10 +1,9 @@
 .App {
   background:rgba(0, 0, 0);
   text-align: center;
-  overflow: hidden;
   height: 100vh;
   margin: 0;
-  position: relative;
+  z-index: 0;
 }
 
 .App-header {
@@ -16,9 +15,20 @@
   font-size: calc(10px + 2vmin);
   color: white;
   position: relative;
+  z-index: 10;
 }
 
-.animation-btn {
+.background-animation {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  transition: opacity 1.5s ease;
+  z-index: 1;
+}
+
+.animation-btn-old {
   background-color: white;
   color: black;
   border: 2px solid black;
@@ -37,26 +47,28 @@
   overflow: hidden; /* Keeps pseudo-elements within the circle */
 }
 
+/* Basic styling for the button */
+.animation-btn {
+  border: 3px solid white;
+  color: white;
+  background-color: rgba(0, 0, 0, 0.4);
+  padding: 10px 20px;
+  font-size: 16px;
+  font-weight: bold;
+  cursor: pointer;
+  outline: none;
+  border-radius: 5px;
+  transition: background-color 0.25s ease, color 0.25s ease;
+}
+
+/* Hover state - changes the background and text color, adds pulse animation */
 .animation-btn:hover {
-  transform: rotate(360deg);
+  background-color: white;
+  color: black;
 }
 
-.animation-btn:before, .animation-btn:after {
-  content: '';
-  position: absolute;
-  width: 100%;
-  height: 2px;
-  /* background-color: black; */
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.animation-btn:after {
-  transform: translate(-50%, -50%) rotate(90deg);
-}
-
+/* Active/click state - compresses the button */
 .animation-btn:active {
-  background-color: #ccc;
   transform: scale(0.95);
+  transition: transform 0.2s ease;
 }
